super-idol/backend/requirements.txt
# Database
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.0

# API
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1

# Authentication
passlib==1.7.4
bcrypt==4.1.2

# Utilities
python-jose==3.3.0
python-multipart==0.0.6

# Flask
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.8.2 